chore: move script files (#632)
