+++ scripts/fn/getExtensionInfos.ts
