+++ scripts/fn/getPackageInfos.ts
