+++ scripts/fn/published-info.ts
