+++ scripts/fn/upload-extensions.ts
